Remember per-state which 'won' metric to use.
Remove State.can_beat_game and move State.update_required_items to Main.
This improves import handling.
